lto: squash order of symbols in partitions
This patch squashes order of symbols in individual partitions, so that
their relative order is conserved, but is not influenced by symbols in
other partitions.
Order of cloned symbols is set to 0. This should be fine because order
specifies order of symbols in input files, which cloned symbols are not
part of.

This is important for incremental LTO because if there is a new symbol,
it otherwise shifts order of all symbols with higher order, which would
diverge them all.

Bootstrapped/regtested on x86_64-pc-linux-gnu

gcc/ChangeLog:

	* lto-cgraph.cc (lto_output_node): Add and use order_remap.
	(lto_output_varpool_node): Likewise.
	(output_symtab): Likewise.
	* lto-streamer-out.cc (produce_asm): Likewise.
	(output_function): Likewise.
	(output_constructor): Likewise.
	(copy_function_or_variable): Likewise.
	(cmp_int): New.
	(lto_output): Generate order_remap.
	* lto-streamer.h (produce_asm): Add order_remap.
	(output_symtab): Likewise.