testsuite: Tweak xfail bogus g++.dg/warn/Wstringop-overflow-4.C:144, PR106120
I added that xfail in February for { ilp32 && c++98_only } and it
looks like it's moved on to lp64 now. :-/  Noted by Rainer
Orth, see the PR.

Tested cris-elf and x86_64-pc-linux-gnu w/wo. -m32.
Ok to commit?

-- >8 --
The conditions under which this this bogus warning is
emitted has changed to not happen for 32-bit targets
anymore.  Adjust accordingly.

	PR testsuite/106120
	* g++.dg/warn/Wstringop-overflow-4.C:144 XFAIL bogus warning for
	lp64 targets with c++98.