Simplify handling of INSN_ and EXPR_LISTs in sched-rgn.cc
This patch simplifies logic behind deps_join(), which will be
important for the upcoming improvements of sched-deps.cc logging.

The only functional change is that when deps_join() is called with
empty state for the 2nd argument, it will not reverse INSN_ and
EXPR_LISTs in the 1st argument.  Before this patch the lists were
reversed due to use of concat_*_LIST().  Now, with copy_*_LIST()
used for this case, the lists will remain in the original order.

gcc/ChangeLog:

	* lists.cc (copy_EXPR_LIST, concat_EXPR_LIST): New functions.
	* rtl.h (copy_EXPR_LIST, concat_EXPR_LIST): Declare.
	* sched-rgn.cc (concat_insn_list, concat_expr_list): New helpers.
	(concat_insn_mem_list): Simplify.
	(deps_join): Update