aarch64: Improve cost of `a ? {-,}1 : b`
While looking into PR 112454, I found the cost for
`(if_then_else (cmp) (const_int 1) (reg))` was being recorded as 8
(or `COSTS_N_INSNS (2)`) but it should have been 4 (or `COSTS_N_INSNS (1)`).
This improves the cost by not adding the cost of `(const_int 1)` to
the total cost.

It does not does not fix PR 112454 as that requires other changes to forwprop
the `(const_int 1)` earlier than combine.

Bootstrapped and tested on aarch64-linux-gnu with no regressions.

gcc/ChangeLog:

	* config/aarch64/aarch64.cc (aarch64_if_then_else_costs):
	Don't add the cost of `1` or `-1`.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>