Fix time-profiler-3.c after r14-5628-g53ba8d669550d3
This testcase started to fail after r14-5628-g53ba8d669550d3 because
IPA-VRP can now start to figure out the functions return a constant
value and there was nothing that profiling needed to profile any more.
This disables IPA-VRP for this testcase to be able to profile again.

Bootrapped/tested on x86_64-linux-gnu with no regressions.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-prof/time-profiler-3.c: Add -fno-ipa-vrp.