ada: Name resolution in expanded instances
In building the tree for an instance of a generic, expansion sets
entity fields on names that refer to things declared outside of the
instance, but leaves the entity field unset on names that should end
up referring to things declared within the instance. These will instead be
set by analysis - the idea is that if a name resolves a certain way in the
generic, then we should get corresponding results if we resolve the
corresponding name in an instance. For this to work, we have to prevent
unrelated declarations that happen to be visible at the point of the
instantiation from participating in resolution. Add code to filter out such
unwanted name resolution candidates.

gcc/ada/

	* sem_ch8.adb (Find_Direct_Name): In the case of a resolving a
	name that occurs within an instantiation, add code to detect and
	filter out unwanted candidate resolutions. The filtering is
	performed via a call to Remove_Interp.

Tested on x86_64-pc-linux-gnu, committed on master.