Fix pr65369.c
There's a caller/callee type mismatch in this test that shows up on
targets where ints are something other than 32 bit types.

Based on reviewing the original bug report, the fix and the part of the
test this fixes, I'm reasonably confident this hasn't compromised the test.

Pushed to the trunk.

Jeff
commit 3da08ffa6df2634092a6292b045568fc326e28e6
Author: Jeff Law <jlaw@ventanamicro.com>
Date:   Sat Dec 2 22:40:41 2023 -0700

    [committed] Fix pr65369.c

    There's a caller/callee type mismatch in this test that shows up on targets
    where ints are something other than 32 bit types.

    Based on reviewing the original bug report, the fix and the part of the test
    this fixes, I'm reasonably confident this hasn't compromised the test.

    gcc/testsuite
            * gcc.c-torture/execute/pr65369.c: Fix type mismatch.