middle-end/112830 - memcpy expansion drops address-spaces
The following makes sure we are not losing address-space info
when expanding a __builtin_memcpy (synthesized by gimplification,
which _might_ be the other actual problem).  The issue is with
get_memory_rtx which is also used by other builtin expansions
but is not aware of address-spaces.  The following fixes that
which resolves an ICE with AVR and properly dispatches to
inline expansion then in emit_block_move_hints.

The alternative, to not intoduce memcpy by gimplification is
possible but IL verification rejects WITH_SIZE_EXPR as part
of GIMPLE assignments, removing that restriction also "works"
(for the testcase).

Alternative is in [2/2].

Sofar untested, will test on x86_64 - in principle I'd prefer 2/2.

Richard.

	PR middle-end/112830
	* builtins.cc (get_memory_rtx): Make address-space aware.

	* gcc.target/avr/pr112830.c: New testcase.