aarch64: Fix up aarch64_print_operand xzr/wzr case
Hi,

This is a v2 of:

https://gcc.gnu.org/pipermail/gcc-patches/2023-November/637612.html

v1 was approved as-is, but this version pulls out the test into a helper
function which is used by later patches in the series.

Bootstrapped/regtested as a series on aarch64-linux-gnu, OK for trunk?

Thanks,
Alex

-- >8 --

This adjusts aarch64_print_operand to recognize zero rtxes in modes other than
VOIDmode.  This allows us to use xzr/wzr for zero vectors, for example.

We extract the test into a helper function, aarch64_const_zero_rtx_p, since this
predicate is needed by later patches.

gcc/ChangeLog:

	* config/aarch64/aarch64-protos.h (aarch64_const_zero_rtx_p): New.
	* config/aarch64/aarch64.cc (aarch64_const_zero_rtx_p): New.
	Use it ...
	(aarch64_print_operand): ... here.  Recognize CONST0_RTXes in
	modes other than VOIDmode.